+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The **third number** is for emergencies when we need to start branches for older
 ### Added
 
 - Official Python 3.12 support.
+  [#9](https://github.com/hynek/stamina/pull/9)
+- Automatic async support.
+  [#10](https://github.com/hynek/stamina/pull/10)
 
 
 ## [22.2.0](https://github.com/hynek/stamina/compare/22.1.0...22.2.0) - 2022-10-06